Implement a QRCode policy

+++ app/Http/Controllers/QRCodeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
 use App\QRCode;
 use App\User;
 use Illuminate\Http\Request;
@@ -14,23 +15,23 @@ class QRCodeController extends Controller
         $this->middleware('auth');
     }
 
-    public function index(User $user) {
-        return $user->codes;
+    public function index() {
+        return Auth::user()->codes;
     }
 
-    public function tempCode(User $user)
+    public function tempCode()
     {
         return QRCodeRenderer::text('Test link')->svg();
     }
 
-    public function new(User $user)
+    public function new()
     {
         request()->validate([
             'name'      => 'required',
             'redirect'  => 'required'
         ]);
 
-        $qrCode = $user->codes()->create([
+        $qrCode = Auth::user()->codes()->create([
             'name'          => request()->name,
             'description'   => request()->description,
             'redirect'      => request()->redirect,
@@ -40,12 +41,14 @@ class QRCodeController extends Controller
         return $qrCode;
     }
 
-    public function update(User $user, QRCode $code) {
+    public function update(QRCode $code) {
+        $this->authorize('update', $code);
         $code->update(request()->all());
         return $code;
     }
 
-    public function delete(USer $user, QRCode $code) {
+    public function delete(QRCode $code) {
+        $this->authorize('update', $code);
         $code->delete();
         return $code;
     }
